+++ pi_Cpp/dataStorage/PirData.h
@@ -28,7 +28,7 @@ class PirData : public Data
 
     /* fetches the full data from MainData. Reduces it to 2 arrays of PLOTRESOLUTION in length
      * Gives back the data in x and y arrays with that length. x (time axis) is cached */
-    uint16_t fetchPirData(uint32_t startT, uint32_t stopT, uint32_t x[],
+    uint16_t fetchPirData(uint32_t startT, uint32_t stopT, double x[],
                           uint16_t y[]);
 
   private:
